pyproject.toml
[project]
name = "creuse-cine-plus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "altair>=5.5.0",
    "matplotlib>=3.10.3",
    "nltk==3.8.1",
    "pandas",
    "plotly-express>=0.4.1",
    "scikit-learn>=1.7.0",
    "streamlit",
    "wordcloud>=1.9.4",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
